+++ src/session/session.js
@@ -192,17 +192,17 @@ module.exports.delHTTPSession = function (req, res, cb) {
 	mem.del(id, cb);
 };
 
-function socketSessionValidation(packet, remoteIp, remotePort, next) {
+function socketSessionValidation(packet, sockType, remoteIp, remotePort, next) {
 	var ce = new gn.lib.CryptoEngine();
 	var res = ce.getSessionIdAndPayload(packet);
-	logger.verbose('validating socket session:', res.sessionId, remoteIp + ':' + remotePort);
+	logger.verbose('validating socket session:', res.sessionId, sockType, remoteIp, remotePort);
 	if (get && set) {
 		logger.verbose('custom getter is defined');
 		get(res.sessionId, function (error, sessionData) {
 			if (error) {
 				return next(error);
 			}
-			_socketSessionValidation(res, remoteIp, remotePort, sessionData, function (error) {
+			_socketSessionValidation(res, sockType, remoteIp, remotePort, sessionData, function (error) {
 				if (error) {
 					return next(error);
 				}
@@ -222,7 +222,7 @@ function socketSessionValidation(packet, remoteIp, remotePort, next) {
 			logger.error('session not found:', res.sessionId);
 			return next(error);
 		}
-		_socketSessionValidation(res, remoteIp, remotePort, sess, function (error) {
+		_socketSessionValidation(res, sockType, remoteIp, remotePort, sess, function (error) {
 			if (error) {
 				return next(error);
 			}
@@ -238,7 +238,7 @@ function socketSessionValidation(packet, remoteIp, remotePort, next) {
 
 // get rid of the code redundancy in secketSessionValidation
 // modifies sess object to be updated
-function _socketSessionValidation(res, remoteIp, remotePort, sess, next) {
+function _socketSessionValidation(res, sockType, remoteIp, remotePort, sess, next) {
 	if (!sess) {
 		logger.error('session not found:', res.sessionId);
 		return next(new Error('SessionNotFound'));
@@ -264,21 +264,26 @@ function _socketSessionValidation(res, remoteIp, remotePort, sess, next) {
 	}
 	// handle initial handshake to keep the client IP and port in the session
 	if (!sess.client) {
+		sess.client = {};
+	}
+	if (!sess.client[sockType]) {
 		logger.info(
 			'handling initial handshake from:',
-			remoteIp + ':' + remotePort,
+			sockType,
+			remoteIp,
 			'session:', res.sessionId
 		);
-		sess.client = {
+		sess.client[sockType] = {
 			ip: remoteIp,
 			port: remotePort
 		};
 	}
 	// check for client IP and port
-	if (sess.client.ip !== remoteIp || sess.client.port !== remotePort) {
+	if (sess.client[sockType].ip !== remoteIp || sess.client[sockType].port !== remotePort) {
 		logger.error(
-			'invalid client IP address/port number detected:',
-			remoteIp + ':' + remotePort,
+			'invalid client IP address detected:',
+			sockType,
+			remoteIp,
 			'session:', res.sessionId
 		);
 		return next('InvalidClient');
@@ -292,7 +297,6 @@ function _socketSessionValidation(res, remoteIp, remotePort, sess, next) {
 }
 
 function socketSessionDecrypt(ce, res, sess, next) {
-
 	var decrypted = ce.decrypt(
 		sess.cipher.cipherKey,
 		sess.cipher.cipherNonce,